requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-monitoring==2.21.0
google-cloud-pubsub==2.21.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
